Exception handler - delete

+++ src/main/java/com/devcaleb/course/services/UserService.java
@@ -2,8 +2,10 @@ package com.devcaleb.course.services;
 
 import com.devcaleb.course.entities.User;
 import com.devcaleb.course.repositories.UserRepository;
+import com.devcaleb.course.services.exceptions.DatabaseException;
 import com.devcaleb.course.services.exceptions.ResourceNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -29,7 +31,15 @@ public class UserService {
     }
 
     public void delete(Long id){
-        userRepository.deleteById(id);
+        if(!userRepository.existsById(id)){
+            throw new ResourceNotFoundException(id);
+        }
+        try {
+            userRepository.deleteById(id);
+        }
+        catch (DataIntegrityViolationException e){
+            throw new DatabaseException("Referential Integrity Failure!");
+        }
     }
 
     public User update(Long id, User obj){
